Refactor to raise error values instead of yielding them.
This code previously contained a weird/brittle pattern of
functions that use `yield` before every possible `error!`,
in order to convey an error code with each error,
as a workaround for the fact that Savi didn't yet
support error values.

Now that [the latest version of Savi](https://github.com/savi-lang/savi/releases/tag/v0.20240529.0)
_does_ support error values, we can use them here.

Note that there is a separate issue uncovered here
for the Savi compiler showing "<unsatisfiable>" errors
when every branch of a choice-inside-a-choice block jumps away.
For now this code does a workaround of adding a fake unreachable
branch that doesn't jump away, for each place this is a problem.